+++ src/main/java/no/ntnu/idatt2106/repository/ListingPictureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.ListingPictureDAO;;
 
-public interface ListingPictureRepository extends JpaRepository<ListingPictureDAO, Long> {
+public interface ListingPictureRepository extends JpaRepository<ListingPictureDAO, Integer> {
     
 }
